add tests, remove some unneeded classes
